Added Open "Source Code" link in a new tab #476
This just makes the links to the source code and Chart.js open in a new
tab. The `rel="noopener noreferrer"` has been added to prevent security
issues in older browsers. This is just an ease of use change that
doesn't affect any of the core speedtest functionality in any way.

Signed-off-by: Matthew Nickson <mnickson@sidingsmedia.com>